+++ src/B0_data_input_json.py
@@ -1,12 +1,18 @@
 import os
 import json
 
+"""
+This module is used to open a json file and parse it as a dict all input parameters for the energy 
+system.
+
+If the user does not give the input parameters "path_input_folder", "path_output_folder" or "path_output_folder_inputs", they are replaced by the default values.
+
+It will be an interface to the EPA.
+"""
+
 
 def load_json(
-    path_input_file,
-    path_input_folder=None,
-    path_output_folder=None,
-    path_input_sequences=None,
+    path_input_file, path_input_folder=None, path_output_folder=None,
 ):
     """Opens and reads json input file and parses it to dict of input parameters.
 
@@ -20,28 +26,13 @@ def load_json(
 
     # The user specified a value
     if path_input_folder is not None:
-        if dict_values["simulation_settings"]["path_input_folder"] != path_input_folder:
-            dict_values["simulation_settings"]["path_input_folder"] = path_input_folder
-    elif path_input_sequences is not None:
-        if (
-            dict_values["simulation_settings"]["path_input_sequences"]
-            != path_input_sequences
-        ):
-            dict_values["simulation_settings"][
-                "path_input_sequences"
-            ] = path_input_sequences
+        dict_values["simulation_settings"]["path_input_folder"] = path_input_folder
 
     # The user specified a value
     if path_output_folder is not None:
-        if (
-            dict_values["simulation_settings"]["path_output_folder"]
-            != path_output_folder
-        ):
-            dict_values["simulation_settings"][
-                "path_output_folder"
-            ] = path_output_folder
-            dict_values["simulation_settings"][
-                "path_output_folder_inputs"
-            ] = os.path.join(path_output_folder, "inputs")
+        dict_values["simulation_settings"]["path_output_folder"] = path_output_folder
+        dict_values["simulation_settings"]["path_output_folder_inputs"] = os.path.join(
+            path_output_folder, "inputs"
+        )
 
     return dict_values